Review 2

+++ cloud-plugin/src/monitoring.rs
@@ -11,10 +11,10 @@ use core::cast;
 #[derive(Debug, Clone)]
 pub struct HostState {
     pub id: ActorId,
-    pub cpu_available: u32,
-    pub memory_available: u32,
-    pub cpu_total: u32,
-    pub memory_total: u32,
+    pub cpu_available: u64,
+    pub memory_available: u64,
+    pub cpu_total: u64,
+    pub memory_total: u64,
 }
 
 #[derive(Debug)]
@@ -24,7 +24,7 @@ pub struct Monitoring {
 }
 
 impl HostState {
-    pub fn new(id: ActorId, cpu_total: u32, memory_total: u32) -> Self {
+    pub fn new(id: ActorId, cpu_total: u64, memory_total: u64) -> Self {
         Self {
             id,
             cpu_available: cpu_total,
@@ -59,7 +59,7 @@ impl Monitoring {
         self.schedulers.insert(scheduler_actor_id.clone());
     }
 
-    pub fn add_host(&mut self, host_id: String, cpu_total: u32, memory_total: u32) {
+    pub fn add_host(&mut self, host_id: String, cpu_total: u64, memory_total: u64) {
         self.host_states.insert(
             host_id.clone(),
             HostState::new(ActorId::from(&host_id), cpu_total, memory_total),
@@ -72,8 +72,8 @@ impl Monitoring {
 #[derive(Debug, Clone)]
 pub struct HostStateUpdate {
     pub host_id: String,
-    pub cpu_available: u32,
-    pub memory_available: u32,
+    pub cpu_available: u64,
+    pub memory_available: u64,
 }
 
 impl Actor for Monitoring {
